% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive_reports.R
\name{archive_reports}
\alias{archive_reports}
\title{Archive reports before a given date}
\usage{
archive_reports(factory = getwd(), before = NULL)
}
\arguments{
\item{factory}{a working directory that contains a valid reportfactory}

\item{before}{a Date or POSIXt object indicating the earliest date keep. All
dates before this date will be archived.}
}
\value{
a named logical vector of archived reports. Successfully archived
  reports will return TRUE and unsuccessfully archived will return FALSE
}
\description{
This places reports in a factory under an \code{_archive} directory,
preserving their original directory structure.
}
