% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_report_sources.R
\name{clean_report_sources}
\alias{clean_report_sources}
\title{Clean the folder report_sources}
\usage{
clean_report_sources(factory = getwd(), quiet = FALSE, remove_cache = TRUE)
}
\arguments{
\item{factory}{the path to a report factory; defaults to the current working
directory}

\item{quiet}{a logical indicating if the function should throw a message if
content gets removed; defaults to `FALSE`}

\item{remove_cache}{a logical indicating if `cache` folder should be removed}
}
\value{
invisibly returns the path to the removed files
}
\description{
This function removes undesirable files
and folders present in `report_sources`
except for:
* `Rmd` files
* `README*` files
* `cache/` (unless `remove_cache` is `FALSE`)
}
\examples{
\dontshow{.old_wd <- setwd(tempdir())}

## create new factory
x <- new_factory()

## check initial content
dir("report_sources", all.files = TRUE)

## add some miscellaneous content
misc_files <- c("toto.txt",
                "#toto.Rmd#",
                "toto.Rmd~",
                "some_file.html",
                ".hidden")
misc_folders <- c("figures",
                  "cache",
                  "outputs_xlsx")

file.create(file.path("report_sources", misc_files))
invisible(
  lapply(file.path("report_sources", misc_folders), dir.create)
)

## check content with all the additional folders and files
dir("report_sources", all.files = TRUE)

## clean the 
clean_report_sources()

## check content after cleanup
dir("report_sources", all.files = TRUE)

\dontshow{setwd(.old_wd); unlink(x)}
}
\author{
Thibaut Jombart
}
