% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_reports.R
\name{compile_reports}
\alias{compile_reports}
\title{Compile an R Markdown report}
\usage{
compile_reports(
  factory = ".",
  reports = NULL,
  params = NULL,
  quiet = TRUE,
  subfolder = NULL,
  timestamp = format(Sys.time(), "\%Y-\%m-\%d_T\%H-\%M-\%S"),
  ...
)
}
\arguments{
\item{factory}{The path to the report factory or a folder within the desired
factory. Defaults to the current directory.}

\item{reports}{Either a regular expression (passed directly to \code{grep()}) that
matches to the report paths you would like to compile or an integer/logical
vector.  If \code{reports} is an integer or character vector then a call of
\code{compile_reports(factory, reports = idx)} is equivalent to
\code{compile_reports(factory, list_reports(factory)[idx])}.}

\item{params}{A named list of parameters to be fed to each report that
matches the regular expression.  If a report has default parameters set in
it's yaml heading and their are duplicates then preference is given to the
value supplied here.}

\item{quiet}{A logical indicating if messages from R Markdown compilation
should be displayed; \code{TRUE} by default.}

\item{subfolder}{Name of subfolder to store results.  Not required but helps
distinguish output if mapping over multiple parameters.}

\item{timestamp}{timevalue used in the folder structure of the report output.
If NULL, \code{format(Sys.time(), "\%Y-\%m-\%d_T\%H-\%M-\%S")} will be used and it's
evaluation forced immediately.}

\item{...}{further arguments passed to \code{rmarkdown::render()}}
}
\description{
Compile an R Markdown report
}
