% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factory_overview.R
\name{factory_overview}
\alias{factory_overview}
\title{Generate an overview of a report factory}
\usage{
factory_overview(path = ".", ...)
}
\arguments{
\item{path}{The path to the report factory or a folder within the desired
factory. Defaults to the current directory.}

\item{...}{Additional arguments passed to \code{fs::dir_tree()}.}
}
\value{
Invisibly returns a character of the files and directories within the
desired folder.
}
\description{
Print contents of directories in a tree-like format.
}
