% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_log.R
\name{filter_log}
\alias{filter_log}
\title{Filter Log}
\usage{
filter_log(log_file, match_exact_type = NULL, most_recent = FALSE,
  outputs_only = FALSE, output_file_types = c(), ...)
}
\arguments{
\item{log_file}{the compile log list (which is located in the root of the 
factory and saved as a .rds file)}

\item{match_exact_type}{a vector of condition types (dots, file, and/or params)
that are to be matched EXACTLY to the log entry (no entries with missing
components of the specified type will be returned)}

\item{most_recent}{a logical indicating if only the last log entry should
be returned; \code{FALSE} by default.}

\item{outputs_only}{a logical indicating if only the outputs of a log entry 
should be returned; \code{FALSE} by default.}

\item{output_file_types}{a vector containing the file types that are to be
returned for each entry}

\item{...}{the arguments that will be used to match and return log entries, 
which should match the structure of the \code{reportfactory::compile_reports}
method}
}
\value{
a list of log entries which match the \code{...} arguments and other
  parameter input
}
\description{
By default the function returns all matches - even if some conditions are 
    missing, or there are other log conditions present
}
\examples{
## This filters a compile log with a set of conditions, which would 
## match the call of the \\code{reportfactory::compile_report} function. 


## new random factory in temp folder
odir <- getwd()
new_factory(tempdir(), include_examples = TRUE)


factory_name = "foo"
report_source_file_name <- list_reports(pattern = factory_name)[1]


## The log entry for a report created with the following arguments:
compile_report(file = report_source_file_name,
               quiet = TRUE, 
               params = list(other = "test"))
               
## The log is saved in the root of the factory
log_file <- readRDS(".compile_log.rds")
               
## Filter the log file for the entry with the same arguments:
filter_log(log_file = log_file,
           file = report_source_file_name,
           quiet = TRUE,
           params = list(other = "test"))
       
setwd(odir)
}
