% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_deps.R
\name{install_deps}
\alias{install_deps}
\title{Install dependencies of reprots within}
\usage{
install_deps(update = FALSE, ...)
}
\arguments{
\item{update}{A logical indicating if packages which are already installed
should be re-installed (\code{TRUE}); otherwise, only missing packages are
installed; defaults to \code{FALSE}.}

\item{...}{Arguments to be passed to \code{install.packages}.}
}
\description{
This function can be used to install package dependencies based of the reports
within the factory.
}
\seealso{
\code{\link{list_deps}} to list dependencies of packages
}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
