% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_reports.R
\name{list_reports}
\alias{list_reports}
\title{List reports in the factory}
\usage{
list_reports()
}
\description{
This function can be used to list available reports stored in the factory, in
inside the \code{report_sources} folder (or any subfolder within).
}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
