% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scripts.R
\name{load_scripts}
\alias{load_scripts}
\title{Load R scripts within factory folder}
\usage{
load_scripts(factory = ".", folder = "scripts", quiet = FALSE)
}
\arguments{
\item{factory}{The path to the report factory or a folder within the desired
factory. Defaults to the current directory.}

\item{folder}{The folder (relative to the factory) from which to source \code{.R}
files; defaults to \verb{scripts/}.}

\item{quiet}{A logical indicating whether messages should be displayed to the
console (\code{TRUE}, default), or not.}
}
\description{
This helper function will load all \code{.R} scripts within a specified factory
and folder. All scripts are loaded within the calling environment and in
alphanumeric order.
}
