% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_alphanum.R
\name{random_alphanum}
\alias{random_alphanum}
\title{Generate 'n' alhapnumeric characters}
\usage{
random_alphanum(n)
}
\arguments{
\item{n}{the number of random alphanumeric characters to use}
}
\description{
Not exported; returns a ranom set of alphanumeric values of length 'n'.
}
