% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_factory.R
\name{random_factory}
\alias{random_factory}
\title{Create a new factory with random name}
\usage{
random_factory(path = getwd(), prefix = "factory", n = 6, ...)
}
\arguments{
\item{path}{the path where the new factory should be created}

\item{prefix}{a `character` indicating the prefix to be used to name the
factory}

\item{n}{the number of random alphanumeric characters to use; defaults to 6.}

\item{...}{further arguments passed to \code{\link{new_factory}}}
}
\description{
This function is a wrapper for \code{\link{new_factory}} which creates a new
factory with a random label, in a specified location. The label itself is
generated by choosing a user-defined number of random alphanumeric
characters.
}
\seealso{
\code{\link{new_factory}} to create new factories with
  user-specified labels.
}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
