% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfh_find_latest.R
\name{rfh_find_latest}
\alias{rfh_find_latest}
\title{Find latest version of a file}
\usage{
rfh_find_latest(pattern, where = getwd(), quiet = FALSE, ...)
}
\arguments{
\item{pattern}{a character string indicating the name of the file to look for (or
regular expression)}

\item{where}{the directory in which to look for the file}

\item{quiet}{a `logical` indicating if messages should be displayed (`TRUE`,
default), or not (`FALSE`)}

\item{...}{further arguments passed to [`list.files`](list.files)

## create random factory and toy files
odir <- getwd()
random_factory(tempdir())
file.create(file.path("data", "linelist_2020-10-01.xlsx"))
file.create(file.path("data", "linelist_2020-10-12.csv"))
file.create(file.path("data", "linelist.xlsx"))
file.create(file.path("data", "contacts.xlsx"))
file.create(file.path("data", "death_linelist_2020-10-13.xlsx"))

## find the latest data with 'linelist' in the name; note that this
## matches both 'linelist' and 'death_linelist' files
rfh_find_latest("linelist")

## same, but this time files starting with 'linelist', i.e. excluding
## 'death_linelist'
rfh_find_latest("^linelist")

## this returns NULL
rfh_find_latest("foobar")

setwd(odir)}
}
\description{
This function will scan recursively a directory, looking for files of a given
name (or matching a given regular expression). Files need to include a date
in a common format (detected automatically). The full path to the most recent
file will be returned.
}
\author{
Thibaut Jombart
}
