% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfh_load_scripts.R
\name{rfh_load_scripts}
\alias{rfh_load_scripts}
\title{Load all R scripts in the factory}
\usage{
rfh_load_scripts(quiet = FALSE)
}
\arguments{
\item{quiet}{A logical indicating whether messages should be displayed to the
console (`TRUE`, default), or not.}
}
\description{
This helper function will load all `.R` scripts within folders `scripts/` and
`src/`. All scripts are loaded using `source()` with the argument `local =
TRUE` so that the current environment (typically that of the compiled report)
will be visible to all sourced scripts.
}
\author{
Thibaut Jombart (thibautjombart@gmail.com)
}
