% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_reports.R
\name{update_reports}
\alias{update_reports}
\title{Compile reports within the factory}
\usage{
update_reports(all = FALSE, quiet = TRUE, ...)
}
\arguments{
\item{all}{a logical indicating if all reports should be compiled; it
defaults to \code{FALSE}, in which case only the latest versions of each
reports will be compuled.}

\item{quiet}{a logical indicating if messages from rmarkdown compilation
should be displayed; \code{FALSE} by default.}

\item{...}{further arguments passed to \code{rmarkdown::render}.}
}
\description{
This function is used to compile several reports stored in the factory,
inside the \code{report_sources} folder (or any subfolder within). Outputs
will be generated in a named and time-stamped directory within
\code{report_outputs}. By default, only the most recent version of each
report is compiled, but an option permits to compile all reports. Reports are
expected to be \code{rmarkdown} documents (\code{.Rmd} extension), named
using the following convention: \code{[name]_[date].Rmd}, where \code{date}
must have the format \code{yyy-mm-dd}, for instance 2018-12-01 for the 1st
December 2018; \code{[name]} should be any explicit name, without blank
spaces or accentuated characters. For instance,
\code{bird_phylogeny_2017-04-01.Rmd} or \code{estimation_R_2018-02-04.Rmd}
are good file names.
}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
