% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_factory.R
\name{validate_factory}
\alias{validate_factory}
\title{Inspect and validate the content of a factory}
\usage{
validate_factory(factory = getwd(), warnings = TRUE, errors = TRUE)
}
\arguments{
\item{factory}{the path to a report factory; defaults to the current working
directory}

\item{warnings}{A logical indicating if warnings should be issued; defaults to \code{TRUE}.}

\item{errors}{A logical indicating if errors should be issued; defaults to \code{TRUE}.}
}
\description{
This function can be used to inspect the content of a factory and make sure
it looks fine. This includes various sanity checks listed in details. The
function returns \code{TRUE} if all checks were done without errors, and
\code{FALSE} otherwise; note that warnings do not count as \code{errors} and
therefore can result in \code{TRUE} being returned.
}
\details{
Checks ran on the factory include:

\itemize{

 \item the directory exists

 \item all mandatory files exist: .here, .gitignore

 \item all mandatory folders exist: report_sources/

 \item all .Rmd reports have unique names once outside their folders (to
 avoid conflicts in outputs)

}
}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
