% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_report_sources.R
\name{clean_report_sources}
\alias{clean_report_sources}
\title{Clean the folder report_sources}
\usage{
clean_report_sources(factory = getwd(), quiet = FALSE,
  remove_cache = TRUE)
}
\arguments{
\item{factory}{the path to a report factory; defaults to the current working
directory}

\item{quiet}{a logical indicating if the function should throw a message if
content gets removed; defaults to `FALSE`}

\item{remove_cache}{a logical indicating if `cache` folder should be kept}
}
\description{
This function removes undesirable files and folders present in `report_sources`
except for:
* `Rmd` files
* `README*` files
* `cache/` (unless `remove_cache` is `FALSE`)
}
\author{
Thibaut Jombart
}
