% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_report_sources.R
\name{clean_report_sources}
\alias{clean_report_sources}
\title{Clean the folder report_sources}
\usage{
clean_report_sources(factory = getwd(), quiet = FALSE, remove_cache = TRUE)
}
\arguments{
\item{factory}{the path to a report factory; defaults to the current working
directory}

\item{quiet}{a logical indicating if the function should throw a message if
content gets removed; defaults to `FALSE`}

\item{remove_cache}{a logical indicating if `cache` folder should be kept}
}
\value{
invisibly returns the path to the removed files
}
\description{
This function removes undesirable files
and folders present in `report_sources`
except for:
* `Rmd` files
* `README*` files
* `cache/` (unless `remove_cache` is `FALSE`)
}
\examples{

## create new random factory in temp folder
odir <- getwd()
x <- random_factory(tempdir())

## add content
dir.create(file.path("report_sources", "_archive"))

## check initial content
dir("report_sources", all.files = TRUE)

## add crap
crap_files <- c("toto.txt",
                "#toto.Rmd#",
                "toto.Rmd~",
                "some_file.html",
                ".hidden")
crap_folders <- c("figures",
                  "cache",
                  "outputs_xlsx")

file.create(file.path("report_sources", crap_files))
sapply(file.path("report_sources", crap_folders), dir.create)

## check content with all crap
dir("report_sources", all.files = TRUE)

## clean the crap
clean_report_sources()

## check content after cleanup
dir("report_sources", all.files = TRUE)

setwd(odir)
unlink(x)
}
\author{
Thibaut Jombart
}
