% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_report.R
\name{compile_report}
\alias{compile_report}
\title{Compile an rmarkdown report}
\usage{
compile_report(file, quiet = FALSE, factory = getwd(),
  clean_report_sources = FALSE, remove_cache = TRUE,
  encoding = "UTF-8", render_params = list(), ...)
}
\arguments{
\item{file}{the full path, or a partial, unambiguous match for the Rmd
report to be compiled.}

\item{quiet}{a logical indicating if messages from rmarkdown compilation
should be displayed; \code{TRUE} by default.}

\item{factory}{the path to a report factory; defaults to the current working
directory}

\item{clean_report_sources}{a logical indicating if indesirable files and
folders in the `report_sources` folder should be cleaned before
compilation; defaults to `FALSE` (no cleanup)}

\item{remove_cache}{a logical passed to `clean_report_sources`, indicating if
the `cache` folder should be considered as undesirable in `report_sources`;
defaults to `TRUE`, in which case `cache` will also be removed if
`clean_report_sources` is `TRUE`}

\item{encoding}{a character string indicating which encoding should be used
when compiling the document; defaults to `"UTF-8"`, which ensures that
non-ascii characters work across different systems}

\item{render_params}{a list that is passed to the `params` argument in
`rmarkdown::render`, which are accessed in the `.Rmd` file with
`params$...`; for instance, if `list(foo = 1:3)` is passed, then the
compilation environment will define an object `params$foo` with value
`1:3`; if defined, the subdirectory name in report-outputs will indicate
the names and first values of `params`; default is an empty list}

\item{...}{further arguments passed to \code{rmarkdown::render}.}
}
\description{
This function can be used to compile a report using its name, or a partial
match for its name. The full path needs not be specified, but reports are
expected to be inside the \code{report_sources} folder (or any subfolder
within). Outputs will be generated in a named and time-stamped directory
within \code{report_outputs}.
}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
