% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_report.R
\name{compile_report}
\alias{compile_report}
\title{Compile an rmarkdown report}
\usage{
compile_report(file, quiet = FALSE, ...)
}
\arguments{
\item{file}{the full path, or a partial, unambiguous match for the Rmd
report to be compiled.}

\item{quiet}{a logical indicating if messages from rmarkdown compilation
should be displayed; \code{TRUE} by default.}

\item{...}{further arguments passed to \code{rmarkdown::render}.}
}
\description{
This function can be used to compile a report using its name, or a partial
match for its name. The full path needs not be specified, but reports are
expected to be inside the \code{report_sources} folder (or any subfolder
within). Outputs will be generated in a named and time-stamped directory
within \code{report_outputs}.
}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
