% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_deps.R
\name{install_deps}
\alias{install_deps}
\title{Install dependencies of reports within}
\usage{
install_deps(factory = ".", update = FALSE, ...)
}
\arguments{
\item{factory}{The path to the report factory or a folder within the desired
factory. Defaults to the current directory.}

\item{update}{A logical indicating if packages which are already installed
should be re-installed (\code{TRUE}); otherwise, only missing packages are
installed; defaults to \code{FALSE}.}

\item{...}{Arguments to be passed to \code{install.packages()}.}
}
\value{
Invisble NULL (called for side effects only).
}
\description{
This function can be used to install package dependencies based on the
reports within the factory.
}
\seealso{
\code{list_deps()} to list dependencies of packages
}
