% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_deps.R
\name{list_deps}
\alias{list_deps}
\title{List dependencies of reports within a factory}
\usage{
list_deps(missing = FALSE, factory = getwd())
}
\arguments{
\item{missing}{A logical indicating if only missing dependencies should be
listed (\code{TRUE}); otherwise, all packages needed in the reports are
listed; defaults to \code{FALSE}}

\item{factory}{the path to a report factory; defaults to the current working
directory}
}
\description{
This function can be used to list package dependencies based of the reports
within the factory. It is a wrapper for \code{checkpoint::scanForPackages}.
}
\seealso{
\code{\link{install_deps}} to install dependencies
}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
