% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_deps.R
\name{list_deps}
\alias{list_deps}
\title{List dependencies of reports within a factory}
\usage{
list_deps(factory = ".", missing = FALSE)
}
\arguments{
\item{factory}{The path to the report factory or a folder within the desired
factory. Defaults to the current directory.}

\item{missing}{A logical indicating if only missing dependencies should be
listed (\code{TRUE}); otherwise, all packages needed in the reports are listed;
defaults to \code{FALSE}}
}
\value{
A character vector of package dependencies.
}
\description{
This function can be used to list package dependencies based of the reports
within the factory. It is a wrapper for \code{checkpoint::scanForPackages}.
}
