% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_outputs.R
\name{list_outputs}
\alias{list_outputs}
\title{List outputs of the factory}
\usage{
list_outputs(factory = getwd())
}
\arguments{
\item{factory}{the path to a report factory; defaults to the current working
directory}
}
\description{
This function can be used to list available report outputs stored in the
factory, in inside the \code{report_outputs} folder (or any subfolder
within).
}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
