% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_outputs.R
\name{list_outputs}
\alias{list_outputs}
\title{List outputs of the factory}
\usage{
list_outputs(factory = ".", pattern = NULL, ...)
}
\arguments{
\item{factory}{The path to the report factory or a folder within the desired
factory. Defaults to the current directory.}

\item{pattern}{An optional regular expression used to look for specific
patterns in report names.}

\item{...}{additional parameters to pass to \code{grep()}}
}
\value{
A character vector containing the names of the reports in the
specified factory (empty if there were no files).
}
\description{
This function can be used to list available report outputs stored in the
factory, in inside the \code{outputs} folder (or any subfolder within).
}
