% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_reports.R
\name{list_reports}
\alias{list_reports}
\title{List reports in the factory}
\usage{
list_reports(factory = ".", pattern = NULL, ...)
}
\arguments{
\item{factory}{The path to the report factory or a folder within the desired
factory. Defaults to the current directory.}

\item{pattern}{An optional regular expression used to look for specific
patterns in report names.}

\item{...}{additional parameters to pass to \code{grep()}}
}
\value{
A character vector containing the names of the reports in the
specified factory (empty if there were no files).
}
\description{
This function can be used to list available reports stored in the factory, in
inside the \code{report_sources} folder (or any subfolder within).
}
