% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_factory.R
\name{new_factory}
\alias{new_factory}
\title{Create a new report factory}
\usage{
new_factory(destination = "new_factory", include_examples = TRUE,
  move_in = FALSE)
}
\arguments{
\item{destination}{the name of the report factory folder to be created}

\item{move_in}{a logical indicating if the current session should move into
the created factory; defaults to \code{FALSE}}

\item{include_example}{a logical indicating if examples of reports shoud be
added to the factory; defaults to \code{TRUE}}
}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
