% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_factory.R
\name{random_factory}
\alias{random_factory}
\title{Create a new factory with random name}
\usage{
random_factory(n = 12, ...)
}
\arguments{
\item{n}{the number of random alphanumeric characters to use; defaults to 12.}

\item{...}{further arguments passed to \code{\link{new_factory}}}
}
\description{
This function is a wrapper for \code{\link{new_factory}} which creates a new
factory with a random label. The label itself is generated by choosing a
user-defined number of random alphanumeric characters.
}
\seealso{
\code{\link{new_factory}} to create new factories with
  user-specified labels.
}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
