% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfh_load_scripts.R
\name{rfh_load_scripts}
\alias{rfh_load_scripts}
\title{Load all R scripts in the factory}
\usage{
rfh_load_scripts(quiet = FALSE)
}
\arguments{
\item{quiet}{A logical indicating whether messages should be displayed to the
console (`TRUE`, default), or not.}
}
\description{
This helper function will load all `.R` scripts within folders `scripts/` and
`src/`. All scripts are loaded using `source()` with the argument `local =
TRUE` so that the current environment (typically that of the compiled report)
will be visible to all sourced scripts.
}
\examples{

## create random factory
odir <- getwd()
random_factory()

## load scripts - nothing to load
rfh_load_scripts()

## create toy script, try again
dir.create("src/")
cat("toto <- 1:10", file = "src/toto.R")
rfh_load_scripts()

## check that toto exists and is 1:10
toto

## restore original working directory
setwd(odir)
}
\author{
Thibaut Jombart (thibautjombart@gmail.com)
}
