% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ship_reports.R
\name{ship_reports}
\alias{ship_reports}
\title{#' Ship Reports}
\usage{
ship_reports(factory = getwd(), match_exact_type = NULL,
  most_recent = TRUE, outputs_only = FALSE, output_file_types = c(),
  ...)
}
\arguments{
\item{factory}{The path the factory directory, defaults to current directory}

\item{match_exact_type}{a vector of condition types passed to `filter_log`
(dots, file, and/or params) that are to be matched EXACTLY to the log entry 
(no entries with missing components of the specified type will be returned)}

\item{most_recent}{a logical indicating passed to `filter_log` if only the 
last log entry should be returned; \code{FALSE} by default.}

\item{outputs_only}{a logical indicating if only the outputs of a log entry 
should be returned passed to `filter_log`; \code{FALSE} by default.}

\item{output_file_types}{a vector containing the file types that are to be
returned for each entry passed to `filter_log`}

\item{...}{the arguments that will be used to match and return log entries, 
which should match the structure of the \code{reportfactory::compile_reports}
method - passed to `filter_log`}
}
\description{
Copies output files for log entries that match given conditions to a dated
`shipped_[YYYY-mm-dd]` folder
}
